Fix consumer behaviour re: extend visibilty and delete
Previously the visibility timeout extender and message deleter were
middlewares that every user of sqsconsumer needed to wire in manually.
The timeout extender had a bug because it was middleware whereby if
previous messages took a long time to process, the visibility timeout
may have expired before the handler started executing, which includes
the visibility timeout extender.

PLAT-5420