Ensure BatchVisibilityExtender lifetime does not outlive parent goroutine (#22)
Heap profiling and analysis has shown that the `startBatchExtender`
method is leaking the ticker created for the `BatchVisibilityExtender`.

It is clear that the ticker will only be stopped if the context expires
before all messages are processed, causing it to potentially be leaked
if any references are held onto after the `startBatchExtender` goroutine
finishes before the context expires.

The `BatchVisibilityExtender` does hold a reference to the ticker in a
goroutine closure, so the ticker cannot be garbage collected until the
goroutine terminates.

The context passed to `BatchVisibilityExtender` is received from the
`Consumer.Run(context.Context)` method, so it would not expire until the
consumer stops.

By creating a child context that is cancelled when the
`startBatchExtender` goroutine returns, the `BatchVisibilityExtender`
internal goroutine cannot outlive the parent goroutine.

These changes to stop the timer and visibility extender goroutine when
the parent goroutine they were created by terminates, should ensure that
these resources are not leaked.